fix test to work on Python 2.7
Trying to call close() on the file handle which was closed by
CloseHandle() seems to cause Python 2.7 to crash.  Python 2.6 and 3.x
do not seem to have this problem however.  Since this is a test case
we're ok with calling flush() instead.  In real life, it's probably
a bug if you try to use Python to close a file handle it's managing
after using CloseHandle to manually manage handle yourself.